final commit - finished mobile version
This is a finish where I did a lot of changes like its background images, positioning of each of the elements in the main and footer. I did some changes to the footer so that I would be able to make the footer close to the preview image as much as possible. Lastly, I create the drop-down sidebar along with transition for easy smooth animation from drop down from the top to show the list of links.